replace icon in components
