+++ src/components/Contacts/ContactDetails/ContactDetailsHeader/ContactDetailsHeader.tsx
@@ -5,7 +5,7 @@ import React from 'react';
 import { useTranslation } from 'react-i18next';
 import theme from '../../../../theme';
 
-import { StarContactIcon } from '../../StarContactIcon/StarContactIcon';
+import { StarContactIconButton } from '../../StarContactIconButton/StarContactIconButton';
 import { useGetContactDetailsHeaderQuery } from './ContactDetailsHeader.generated';
 import { ContactHeaderAddressSection } from './ContactHeaderSection/ContactHeaderAddressSection';
 import { ContactHeaderPhoneSection } from './ContactHeaderSection/ContactHeaderPhoneSection';
@@ -101,9 +101,10 @@ export const ContactDetailsHeader: React.FC<Props> = ({
           ) : null}
         </HeaderBarContactWrap>
         <HeaderBarButtonsWrap>
-          <ButtonWrap>
-            <StarContactIcon hasStar={false} />
-          </ButtonWrap>
+          <StarContactIconButton
+            accountListId={accountListId}
+            contactId={contactId}
+          />
           <ButtonWrap>
             <MoreButtonIcon />
           </ButtonWrap>
